User classes moved to users directory
