+++ users/class_doctor.py
